fsd/fsd/requirements.txt
numpy==1.26.3
torch==2.0.1
torchvision==0.15.2
opencv-python==4.7.0.72
matplotlib==3.8.3
einops==0.7.0
av==11.0.0
imageio==2.34.0
hydra-core==1.3.2
tqdm>=4.65.0
apng==0.3.4
Pillow==9.4.0
scipy==1.12.0
mypy==1.8.0
fsspec==2024.2.0
gcsfs==2024.2.0
pandas>=1.1.5
skia-python==87.5
lpips==0.1.4
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git@6fdee8f2727f4506cfbbe553e23b895e27956588
easydict==1.12
datasets==2.20.0
